// src/api-client/models/MonthlyTransactionType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Budget Manager API
 * Personal finance management API for tracking income, expenses, accounts, and budgets.
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const MonthlyTransactionType = {
    NUMBER_0: 0,
    NUMBER_1: 1
} as const;
export type MonthlyTransactionType = typeof MonthlyTransactionType[keyof typeof MonthlyTransactionType];


export function instanceOfMonthlyTransactionType(value: any): boolean {
    for (const key in MonthlyTransactionType) {
        if (Object.prototype.hasOwnProperty.call(MonthlyTransactionType, key)) {
            if (MonthlyTransactionType[key as keyof typeof MonthlyTransactionType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function MonthlyTransactionTypeFromJSON(json: any): MonthlyTransactionType {
    return MonthlyTransactionTypeFromJSONTyped(json, false);
}

export function MonthlyTransactionTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): MonthlyTransactionType {
    return json as MonthlyTransactionType;
}

export function MonthlyTransactionTypeToJSON(value?: MonthlyTransactionType | null): any {
    return value as any;
}

export function MonthlyTransactionTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): MonthlyTransactionType {
    return value as MonthlyTransactionType;
}